medvak-agent-suite/bot/pyproject.toml
[project]
name = "medvak-bot"
version = "0.1.0"
description = "Telegram-бот: PREVIEW CSV → запись в NocoDB через агент"
requires-python = ">=3.11"

dependencies = [
  "python-telegram-bot==21.*",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
]

[tool.pyright]
typeCheckingMode = "basic"
